+++ source/ftrack/exception.py
@@ -4,16 +4,21 @@
 import sys
 import traceback
 
+import ftrack.entity.base
+
 
 class Error(Exception):
     '''ftrack specific error.'''
 
     default_message = 'Unspecified error occurred.'
 
-    def __init__(self, message=None, details=None, **kw):
+    def __init__(self, message=None, details=None):
         '''Initialise exception with *message*.
 
         If *message* is None, the class 'default_message' will be used.
+
+        *details* should be a mapping of extra information that can be used in
+        the message and also to provide more context.
 
         '''
         if message is None:
@@ -21,12 +26,15 @@ class Error(Exception):
 
         self.message = message
         self.details = details
+        if self.details is None:
+            self.details = {}
+
         self.traceback = traceback.format_exc()
 
     def __str__(self):
         '''Return string representation.'''
         keys = {}
-        for key, value in self.__dict__.iteritems():
+        for key, value in self.details.iteritems():
             if isinstance(value, unicode):
                 value = value.encode(sys.getfilesystemencoding())
             keys[key] = value
@@ -71,7 +79,9 @@ class UnrecognisedEntityTypeError(EntityTypeError):
 
     def __init__(self, entity_type, **kw):
         '''Initialise with *entity_type* that is unrecognised.'''
-        self.entity_type = entity_type
+        kw.setdefault('details', {}).update(dict(
+            entity_type=entity_type
+        ))
         super(UnrecognisedEntityTypeError, self).__init__(**kw)
 
 
@@ -91,9 +101,11 @@ class InvalidStateTransitionError(InvalidStateError):
 
     def __init__(self, current_state, target_state, entity, **kw):
         '''Initialise error.'''
-        self.current_state = current_state
-        self.target_state = target_state
-        self.entity = entity
+        kw.setdefault('details', {}).update(dict(
+            current_state=current_state,
+            target_state=target_state,
+            entity=entity
+        ))
         super(InvalidStateTransitionError, self).__init__(**kw)
 
 
@@ -112,7 +124,9 @@ class ImmutableAttributeError(AttributeError):
 
     def __init__(self, attribute, **kw):
         '''Initialise error.'''
-        self.attribute = attribute
+        kw.setdefault('details', {}).update(dict(
+            attribute=attribute
+        ))
         super(ImmutableAttributeError, self).__init__(**kw)
 
 
@@ -123,7 +137,9 @@ class CollectionError(Error):
 
     def __init__(self, collection, **kw):
         '''Initialise error.'''
-        self.collection = collection
+        kw.setdefault('details', {}).update(dict(
+            collection=collection
+        ))
         super(CollectionError, self).__init__(**kw)
 
 
@@ -144,7 +160,9 @@ class DuplicateItemInCollectionError(CollectionError):
 
     def __init__(self, item, collection, **kw):
         '''Initialise error.'''
-        self.item = item
+        kw.setdefault('details', {}).update(dict(
+            item=item
+        ))
         super(DuplicateItemInCollectionError, self).__init__(collection, **kw)
 
 
@@ -170,3 +188,148 @@ class EventHubPacketError(EventHubError):
     '''Raise when event hub encounters an issue with a packet.'''
 
     default_message = 'Invalid packet.'
+
+
+class PermissionDeniedError(Error):
+    '''Raise when permission is denied.'''
+
+    default_message = 'Permission denied.'
+
+
+class LocationError(Error):
+    '''Base for errors associated with locations.'''
+
+    default_message = 'Unspecified location error'
+
+
+class ComponentNotInAnyLocationError(LocationError):
+    '''Raise when component not available in any location.'''
+
+    default_message = 'Component not available in any location.'
+
+
+class ComponentNotInLocationError(LocationError):
+    '''Raise when component(s) not in location.'''
+
+    default_message = (
+        'Component(s) {components} not found in location {location}.'
+    )
+
+    def __init__(self, components, location, **kw):
+        '''Initialise with *components* and *location*.'''
+        if isinstance(components, ftrack.entity.base.Entity):
+            components = [components]
+
+        kw.setdefault('details', {}).update(dict(
+            components=', '.join([str(component) for component in components]),
+            location=location
+        ))
+        super(ComponentNotInLocationError, self).__init__(**kw)
+
+
+class ComponentInLocationError(LocationError):
+    '''Raise when component already exists in location.'''
+
+    default_message = (
+        'Component {component} already exists in location {location}.'
+    )
+
+    def __init__(self, component, location, **kw):
+        '''Initialise with *component* and *location*.'''
+        kw.setdefault('details', {}).update(dict(
+            component=component,
+            location=location
+        ))
+        super(ComponentInLocationError, self).__init__(**kw)
+
+
+class AccessorError(Error):
+    '''Base for errors associated with accessors.'''
+
+    default_message = 'Unspecified accessor error'
+
+
+class AccessorOperationFailedError(AccessorError):
+    '''Base for failed operations on accessors.'''
+
+    default_message = 'Operation {operation} failed: {error}'
+
+    def __init__(self, operation='', resource_identifier=None, **kw):
+        kw.setdefault('details', {}).update(dict(
+            operation=operation,
+            resource_identifier=resource_identifier
+        ))
+        super(AccessorOperationFailedError, self).__init__(**kw)
+
+
+class AccessorUnsupportedOperationError(AccessorOperationFailedError):
+    '''Raise when operation is unsupported.'''
+
+    default_message = 'Operation {operation} unsupported.'
+
+
+class AccessorPermissionDeniedError(AccessorOperationFailedError):
+    '''Raise when permission denied.'''
+
+    default_message = (
+        'Cannot {operation} {resource_identifier}. Permission denied.'
+    )
+
+
+class AccessorResourceIdentifierError(AccessorError):
+    '''Raise when a error related to a resource_identifier occurs.'''
+
+    default_message = 'Resource identifier is invalid: {resource_identifier}.'
+
+    def __init__(self, resource_identifier, **kw):
+        kw.setdefault('details', {}).update(dict(
+            resource_identifier=resource_identifier
+        ))
+        super(AccessorResourceIdentifierError, self).__init__(**kw)
+
+
+class AccessorFilesystemPathError(AccessorResourceIdentifierError):
+    '''Raise when a error related to an accessor filesystem path occurs.'''
+
+    default_message = (
+        'Could not determine filesystem path from resource identifier: '
+        '{resource_identifier}.'
+    )
+
+
+class AccessorResourceError(AccessorError):
+    '''Base for errors associated with specific resource.'''
+
+    default_message = 'Unspecified resource error: {resource_identifier}'
+
+    def __init__(self, operation='', resource_identifier=None, error=None,
+                 **kw):
+        kw.setdefault('details', {}).update(dict(
+            operation=operation,
+            resource_identifier=resource_identifier
+        ))
+        super(AccessorResourceError, self).__init__(**kw)
+
+
+class AccessorResourceNotFoundError(AccessorResourceError):
+    '''Raise when a required resource is not found.'''
+
+    default_message = 'Resource not found: {resource_identifier}'
+
+
+class AccessorParentResourceNotFoundError(AccessorResourceError):
+    '''Raise when a parent resource (such as directory) is not found.'''
+
+    default_message = 'Parent resource is missing: {resource_identifier}'
+
+
+class AccessorResourceInvalidError(AccessorResourceError):
+    '''Raise when a resource is not the right type.'''
+
+    default_message = 'Resource invalid: {resource_identifier}'
+
+
+class AccessorContainerNotEmptyError(AccessorResourceError):
+    '''Raise when container is not empty.'''
+
+    default_message = 'Container is not empty: {resource_identifier}'